# TP3/api_metacritic_game/openapi.yaml
openapi: 3.0.1
info:
  description: "Projet Final"
  version: "1.0.6"
  title: "MetaCritic"
...
paths:
  /users:
    get:
      tags:
        - Users
      summary: All Users
      description: Returns all users
      operationId: "getUsers" 
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Users'
    
    post:
      tags:
        - Users
      summary: Add user
      description: Add user
      operationId: "addUser" 
      requestBody:
        $ref: '#/components/requestBodies/User'
      responses:
        '201':
          description: criticsEditors response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Users'

    put:
      tags:
        - Users
      summary: Update User
      description: Update a user by his id
      operationId: "updateUser"
      requestBody:
        $ref: '#/components/requestBodies/User'
      responses:
        '201':
          description: users response
  /users/{id}:        
    get:
      tags:
        - Users
      summary: Users By Id
      description: Returns a user by id
      operationId: "getUserById" 
      parameters:
        - $ref: '#/components/parameters/idInPath'
      responses:
        '200':
          description: users response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Users'
                 
    delete:
      tags:
        - Users
      summary: Delete User By Id
      description: Delete a user by his id
      operationId: "DeleteUserById" 
      parameters:
        - $ref: '#/components/parameters/idInPath'
      responses:
        '204':
          description: users response
  /users/role/{id}:
    put:
      tags:
        - Users
      summary: Update role's user by id
      description: Returns new role and id
      operationId: "updateRoleByUserId" 
      parameters:
        - $ref: '#/components/parameters/idInPath'
      responses:
        '200':
          description: users response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Users'
...
  /reviews:
    get:
      tags:
        - Reviews
      summary: All Reviews
      description: Returns all the reviews mayde by editors
      operationId: "getReviews" 
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Reviews'

    post:
      tags:
        - Reviews
      summary: Add Review
      description: Add review
      operationId: "addReview" 
      requestBody:
        $ref: '#/components/requestBodies/Review'
      responses:
        '201':
          description: review response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Reviews'

    put:
      tags:
        - Reviews
      summary: Update Review
      description: Update a review by his id
      operationId: "updateReview" 
      requestBody:
        $ref: '#/components/requestBodies/Review'
      responses:
        '201':
          description: review response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Reviews'
  /reviews/{id}:
    get:
      tags:
        - Reviews
      summary: Reviews By Id
      description: Returns a review by his id
      operationId: "getreviewById" 
      parameters:
        - $ref: '#/components/parameters/idInPath'
      responses:
        '200':
          description: review response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Reviews'

    delete:
      tags:
        - Reviews
      summary: Delete Review By Id
      description: Delete a review by his id
      operationId: "DeleteReviewById" 
      parameters:
        - $ref: '#/components/parameters/idInPath'
      responses:
        '204':
          description: review response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Reviews'
...
components:
  parameters:
    idInPath:
      required: true
      name: id
      in: path
      schema:
        type: integer
...
  requestBodies:
    User:
      description: A JSON object containing an user
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Users'
          examples:
            userExample:
              $ref: '#/components/examples/Users'
...
    Review:
      description: A JSON object containing a review
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/examples/Reviews'
          examples:
            reviewsExample:
              $ref: '#/components/examples/Reviews'
...
  schemas:
    Users:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        pseudo:
          type: string
          example: "user_1"
        email:
          type: string
          example: "user_1@gmail.com"
...
    Reviews:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        avg:
          type: number 
          format: int64
          example: 9.8   
        description:
          type: string
          example: "Take on the role of James Bond as played by Daniel Craig in all new SKYFALL levels. Unlock three new single player levels following the storyline of the SKYFALL movie. Levels take place in Istanbul and Shanghai as the player (Bond) attempts to track down "
        title:
          type: string
          example: "007 Legends: Skyfall"
        release:
          type: string
          example: "11/02/2002"
        idCriticEditor:
          type: integer
          format: int64
          example: 1
...
  examples:
    Reviews:
      summary: An example of a review
      value:
        id: 1
        avg: 15
        description: "très bon jeu"
        title: "Persona 5"
        release: "11-02-2002"
        idCriticEditor: 2
...
    Users:
      summary: An example of a user
      value:
        id: 3
        pseudo: "johnDoe"
        email: "johnDoe@gmail.com"
        hashedPassword: "$2b$10$t7oxiwchWGHa/B9w0AzrYO2WH2rQbA86YSuQjSTmwIrpC/0ZXN7V2"
        roles: "user"
...